Sanity System
The player will now hear specific sounds and music while under maximum sanity additionally the logic to spawn in a unique enemy is implemented for reaching maximum sanity as well. The players sanity also affects the dialogue towards NPC's, for now this is all the functionality that I want for the sanity system.